chore(workspace): extract workspace initialize domain logic
wip: simplify type

wip: reference error type

wip: extract initialize workspace

wip: extract workspace type
