+++ src/kairos/workspace/adapter/http/create-workspace/create-workspace-response.ts
@@ -1,7 +1,9 @@
 import { type ResultHttpMapping } from '@/interface/http/result-to-http';
 import { HTTP_CONFLICT, HTTP_CREATED } from '@/interface/http/status-code';
 import type { WorkspaceSlugConflictError } from '@/kairos/workspace/domain/errors';
-import type { Workspace } from '@/kairos/workspace/domain/workspace';
+import { workspaceSlugConflictError } from '@/kairos/workspace/domain/errors';
+
+import { Workspace } from '@/kairos/workspace/domain/workspace.type';
 
 export const createWorkspaceResponse: ResultHttpMapping<Workspace, WorkspaceSlugConflictError> = {
   success: {
@@ -9,7 +11,7 @@ export const createWorkspaceResponse: ResultHttpMapping<Workspace, WorkspaceSlug
   },
   error: {
     byType: {
-      WORKSPACE_SLUG_CONFLICT: {
+      [workspaceSlugConflictError.type]: {
         status: HTTP_CONFLICT,
       },
     },
