+++ src/kairos/workspace/create-workspace/insert-workspace.type.ts
@@ -1,6 +1,7 @@
 import type { Result } from '@/kairos/shared/result/result.type';
 import type { WorkspaceSlugConflictError } from '@/kairos/workspace/domain/errors';
-import type { InitialWorkspace, Workspace } from '@/kairos/workspace/domain/workspace';
+import { InitialWorkspace } from '@/kairos/workspace/domain/initialize-workspace';
+import { Workspace } from '@/kairos/workspace/domain/workspace.type';
 
 type InsertWorkspaceResult = Result<Workspace, WorkspaceSlugConflictError>;
