+++ src/kairos/workspace/domain/workspace.type.ts
@@ -1,0 +1,6 @@
+export type Workspace = Readonly<{
+  id: string;
+  environments: string[];
+  name: string;
+  slug: string;
+}>;
